requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==35.0.0
defusedxml==0.7.1
Django==3.1.7
django-cors-headers==3.7.0
djangorestframework==3.12.1
djangorestframework-datatables==0.6.0
idna==3.2
oauthlib==3.1.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
